Cleanup api for visitQuantileEstimator
Summary:
This api takes a fmt argument and a prefix that is passed to the fmt argument
and expects a suffix to construct the name. There is already an assumption
that the fmt argument is {}_{}. Instead we now jsut take in the prefix and
construct the proper name. This should not cause any functional change.

Also, use this api uniformly in all percentile stats reporting directly.

Reviewed By: therealgymmy

Differential Revision: D26739375

fbshipit-source-id: 07e625be120a3a1197fd03c4dd32df459c557a7b