TimeStampTicker -- create ticker according to time stamp.
Summary:
This diff creates a ticker according to a time stamp passed in by the application.

The ticker syncs the threads at bucket, which is configurable at 3600 ticks. So when passed in timestamps by seconds, the ticker syncs the threads every hour.

A thread attempting to update a time stamp into the next bucket will wait until all other threads finish the current bucket.

Reviewed By: therealgymmy

Differential Revision: D26211306

fbshipit-source-id: af3fb0ce3ad71d8b31ef011fb1bbdd5adacc69a6