minor cleanup for virtual specification
Summary:
only acceptImpl, getCountersImpl are intended to be overridden by
implementations. The rest need not be virtual and marked as final in the base
class since we don't intend to override them. Also when we override, we don't
need to mention it as virtual.

there should be no functional change.

Reviewed By: haowu14

Differential Revision: D26773326

fbshipit-source-id: 9628992a3ab6f915ded0ac1b9ab0caf381060c40