use Item::Key instead of std::string
Summary:
cachelib users interact with cachelib apis through a key that is represented
using Item::Key. Use that instead of introducing a new type in public api.
This will enable reusing some of the functionality in places where we have
item key instead of std::string. If this was std::string, callers would be
forced to construct a std::string to confirm to this api

Reviewed By: therealgymmy

Differential Revision: D26773481

fbshipit-source-id: 2716c6db4497c407a0e045e9fa51ac15beedcd34