Create NavyConfig class to change Navy options into a declarative way
Summary:
**Context**
Navy's options are passed to NvmCache as a folly::dynamic (it is a map): https://fburl.com/diffusion/ta6lvapa, where user can set whatever keys for the config option and change it anywhere, which makes the codebase messy and hard to manage.

Now we want to replace this folly::dynamic way by a new class `NavyConfig` . In this class, we
- set private fields based on https://fburl.com/diffusion/a9t60d0c
- create getters/setters for all fields
- create serialization and other useful member functions

In this diff, the getters and setters are in the more basic/dummy version. In the next diff, these APIs will be modified based on the dependency of each setting.

Reviewed By: sathyaphoenix

Differential Revision: D26700700

fbshipit-source-id: 7164c2537217b7f06a966ace817d1f9d52da5cdd