Move the suppression earlier in the file.
Summary:
TSAN still complained on the earlier load in incRef and decRef. So moving the suppression earlier.

Also, I noticed that the earlier load was not an atomic load which is actually invalid, so I changed that to __atomic_load_n. This isn't to satisfy TSAN (as TSAN is already suppressed anyway), but just for code correctness.

Reviewed By: therealgymmy

Differential Revision: D27100664

fbshipit-source-id: 37e1f8ac379af6de1b2f1a31ee6af0320c7c5e52