Add enabled flag in NavyConfig
Summary:
Migration plan doc: https://fb.quip.com/yBNAAr9Ou8lb

**Why we need this field?**
This boolean field is only for the sake of migration. During migration, there will be middle point where some use cases set data in `NavyConfig` while the others don't; therefore, we need a flag to differentiate whether we can use `NavyConfig` getters to get data.

**When the field will be set?**
It is set to be `true` in every `NavyConfig` setter. The default value is `false`.

**Clean up**
When the migration is fully finished, the field is no longer useful and need to be removed.

Client Impact: None

Reviewed By: haowu14

Differential Revision: D27248223

fbshipit-source-id: 46529f1760d5b7723834e1cef26b68276c7c75f7