Fix stats bug in Navy device bytes written
Summary:
**Client Impact:** On ODS, navy's `navy_device_bytes_written` should show up correctly.

 ---

The write stat is incorrect because we bump it by the whole size of the writes each time a write succeeds. This is a problem because in Device, we have a concept of "max write", where we will split any writes larger than that into several writes. This means this stat was bumped incorrect and it was equivalent to "write-size * (write-size / max-write-size)".

Reviewed By: leozzx

Differential Revision: D27307872

fbshipit-source-id: 77ed7ee7ff7d462d87fc1c8b9eef35e91a94c130