Avoid repeated calls to std::time.
Summary:
5%+ of the time in memcached is spent calling isExpired, the only
complicated piece of which is a call to std::time(nullptr). This
diff avoids repeated calls to std::time by caching the current time
once and feeding that to the isExpired function.

The code can perform a more expensive check that involves obtaining
a handle and running more checks, one of which is another call to
isExpired(). If this reveals that an item should have been expired,
we can update our current time.

Reviewed By: sathyaphoenix

Differential Revision: D27143738

fbshipit-source-id: b93c3ea80f900f2a8e5d14c2bd7631a7345236e7