Modify admission policy setters in NavyConfig
Summary:
We created naive setters for `NavyConfig` in D26700700 (https://github.com/facebookincubator/CacheLib/commit/98fa37539055de5018595ef06b3ad6a66ef0163e). Now we want to modify the setters by combining relevant parameters, throwing exceptions for invalid input, etc.

In this series of diffs, we make the change for each of the following sections:
1. AP settings
2. device settings
3. BlockCache settings
4. BigHash settings
5. Job scheduler and other settings

In this diff, we modify the setters for AP settings:
- `admissionProbability` should not be set when it is "dynamic_random" policy
- `admissionWriteRate`, `maxWriteRate`, `admissionProbBaseSize` should not be set when it is `random` policy
- check invalid input

Reviewed By: haowu14

Differential Revision: D27206151

fbshipit-source-id: 054650bd0c27524df513c967d7788e3e78541ed1