Fast negative lookup for NvmCache
Summary:
Navy maintains state in DRAM for bighash and block cache that can be used to
quickly peek and confirm if IO is necessary for lookups. In case of misses,
most of these can be resolved without going through the async path. Async path
involves overheads like creation of baton, shared_ptrs, thread hop and
spinning. In addition, this will improve the nvm lookup latency from 100us to
~2us.

To ensure correctness, we see if there are any inflight put with a context for
the same shard. Put contexts are now created under the fill lock.

Reviewed By: therealgymmy

Differential Revision: D20812373

fbshipit-source-id: acbbcf3839ab935e8812b271a44295494289aa04