Add micro benchmark for cachelib operations
Summary:
**Client Impact:** none
------

Goal of this is to establish a new baseline on the latency of various cachelib operations. This can serve as a starting point to understand the operation cost and pinpoint where we can improve further.

For read benchmarks, I have added some writes but I excluded counting them as part of the average lookup time. For write benchmarks, I exclusively measured writes to see how expensive allocation is. Allocating (evicting) is 20ms across 16 threads for a single alloc size due to contention on a single LRU (which is the case for AdRanker in production).

There're other smaller costs. E.g. generating a random number is 20ns. Reading a string is 40ns.

Finally, for lookups, we're heavily affected by cache locality. The "peek" latency range from ~200ns for 10K objects to ~400ns for 1M objects. Peek operation only checks the hashtable so this is mostly the latency accessing hashtable and also reading the item's key for a string comparison.

Differential Revision: D27317904

fbshipit-source-id: 8bf64e58865b846010fb85aabc80932dd2824a7a