Benchmark a ticker based on steady clock
Summary:
To support DynamicRandomAP in simulator, we need to make it work with tickers.
It uses steady_clock as the time. This diff benchmarks a ticker based on steady_clock. The benchmark result isn't good. So we can't just directly use a ticker pointer.

Later in DynamicRandomAP will have a shared_ptr of the ticker. If the ticker pointer is null, the inline function will be used, otherwise (in cachebench simulator) the ticker is used.

Client Impact: None.

Reviewed By: therealgymmy

Differential Revision: D27279005

fbshipit-source-id: 9206cc5f67b0dc592f628881d0f1b1f669950672