Pass in CacheConfig& to make nvmadmission policy
Summary:
Even after the test config is fixed, there is still exception thrown. The reason is that the ticker around line 83 of fbDep.cpp is captured as a nullptr. I don't know why it didn't fail the check in the past. I thought it was because smart pointers are too smart.

We discussed about using CacheConfig& as the parameter and return of `customizeCacheConfigForFacebook`. I didn't end up doing that because:
1. It is hard to guarantee that the reference returned would be the same reference on which config.nvmAdmissionPolicy is called. Every engineer who contributes to cachebench needs to know this in order to not make a copy and create a new instance and lose the reference.
2. Even if we somehow enforce that it should be the same reference, it is still hard because `CacheConfig` is part of `CacheBenchConfig` when calling `customizeCacheConfigForFacebook`. If we want to maintain the same reference, we would need to move it out of `CacheBenchConfig` and I don't think that's much cleaner.

The change in this diff is to have `nvmAdmissionPolicyFactory` take a `CacheConfig` reference. In this way, we can make sure the ticker pointer is present when this function is called.

Reviewed By: therealgymmy

Differential Revision: D27759830

fbshipit-source-id: b916b76d170c5c8a2faf6be7296eead58da94b5e