Use atomics for updating lruRefreshTime
Summary:
**Client Impact:** None

-----

Use atomics to annotate our intent clearly. We mark it with std::memory_order_relaxed, which should produce the same code as previously on intel processors.

To explain in more details, we have a concept called "lruRefreshTime" in LRU (and other eviction policies) which governs how often we will "promote" an item to the head on access. Every so often, we will update this "lruRefreshTime" while other threads could be reading it. However, we don't need the reader threads to definitely see the latest value. It is totally okay that a reader sees an old "lruRefreshTime".

Differential Revision: D27832928

fbshipit-source-id: 4187ed81590a5c4d04a07c95ae28b3cc5ea7cd00