Migrate cachebench from dipperOptions to navyConfig
Summary:
**Context:**
We are going to deprecate `dipperOptions` and use `NavyConfig`, a new class we recently created, to set Navy related parameters for `nvmCache`.

**What is NavyConfig?**
`NavyConfig` is a more declarative way to do the setup. It contains all the same fields as in `dipperOptions`. However, we need to use setters to set these fields instead of directly assigning the values. We have all the getters and setters ready for use, please check https://fburl.com/diffusion/lero4w0v for more details.

**Migration progress:**
We already added the logic in `nvmCache` to read data from `NavyConfig` https://fburl.com/diffusion/zlw6j2rc.
Now we are in the stage of migrating all the use cases to set parameters in `NavyConfig`.
Note: NavyConfig will be "enabled" when any of the setters are called.

**This diff:**
Migrate cachebench from using dipperOptions to using NavyConfig.
Note: old fields that we did not set through `NavyConfig`:
```
- nvmConfig.dipperOptions["dipper_force_reinit"] = true;
- nvmConfig.dipperOptions["dipper_navy_read_buffer"] = config_.dipperNavyStackAllocReadBufSizeKB * 1024;

```

Reviewed By: leozzx

Differential Revision: D27364714

fbshipit-source-id: 1506dde010463f4e0fb90475dfeefa1027ffe974