Fix a dead loop in PieceWiseReplayGenerator
Summary:
Currently cachebench may hang. If we fail to create CacheStressor, it will destruct its member PieceWiseReplayGenerator, which will wait for the thread to read trace to be done. However, since no one remove traces from active queue, the thread to read traces to queue may be blocked.
Fix this by marking PieceWiseReplayGenerator as done in destructor to let the thread stop reading traces and exist.

Reviewed By: therealgymmy

Differential Revision: D28581824

fbshipit-source-id: e2d5652d9c238a8b22c7372ff4e6406a84c14ff6