Allow different units of timestamp in replay generator
Summary:
While bigcache use traces at milliseconds, TAO use seconds as unit in timestamp.

In this diff, we decide whether to divide the timestamp by 1000 by whether the timestamp is greater than the current time. This is relatively save because it is based on the assumption that the simulation timestamps are always small then the current time.

To be more rigorous, we could keep dividing 1000 until the timestamp is smaller than current. But we don't have any use case beyond milliseconds.

Reviewed By: tang-jianfeng

Differential Revision: D27756813

fbshipit-source-id: e53c7cad31253ed847eff9c473ca925f1fe0098e