Change default max write rate for DynamicRandom to 160MB/s
Summary:
**Client Impact:** None. Existing users of DynamicRandom still have the same max write rate

 ---

Most HybridCache use-cases are on T10 with an endurance limit of around 120MB/s. Our default max rate was 80MB/s which is lower than the limit. The goal of DynamicRandom is to write less during peak and write more during trough. So this diff changes the default max rate to 160MB/s to allow some leeway to write above endurance limit if we wrote less than that during trough.

Reviewed By: sathyaphoenix, mogeb

Differential Revision: D28884546

fbshipit-source-id: dd24ee5abe2e6e8091b15cb1d2fa05e02584bc31