temporary test CacheBench in Servicelab Benchmark
Summary:
Test CacheBench in Servicelab Benchmark, this is a temporary test to see how cachebench works with servicelab.
Ideally we can have a single build rule for cachebench, and several servicelab standalone rules with different config and name.
We also need to render all metric we'd like to compare for benchmarking.

Reviewed By: therealgymmy

Differential Revision: D28908842

fbshipit-source-id: f194c3f00f1c874d3d0bb652d46d64838ecbceb8