Update wrapAsIOBuf to use item size as capacity
Summary:
**Client Impact**: None
-----
There's no need for us to pass in the beginning of an item and then later on adjust its size. We simply need to pass in the pointer pointing at Item::getMemory() and the user-requested size. This is fine as we don't need access to the "item pointer" to perform cleanup as wrapIOBuf does NOT claim ownership.

Reviewed By: sathyaphoenix

Differential Revision: D28813918

fbshipit-source-id: a4f50236257355dfd876cd6f327225e7268f94ad