Check if key could exist before scheduling async remove job
Summary:
Check if key could exist before scheduling async remove job

When testing with flat memcache with no items in flash, we found that about 5% of cpu time is taken by NvmCache::remove() - https://fburl.com/cv/gorzaljx. Since there are no items on flash, it is unnecessary to schedule a job to remove from flash.

This change checks to see if the item could exist on flash before scheduling the remove job

Reviewed By: leozzx

Differential Revision: D29071883

fbshipit-source-id: ef8da78339cd178c6e64fceab6cb909dd28f9c8d