cachebench@servicelab
Summary:
Integrate CacheBench with Servicelab benchmark tool, it will be triggered automatically by any cachelib diff. We can enable periodically and land-time run in the future when this is stable.

Different than cogwheel test, this one focus on performance (latency and throughput), tests with consistency check or other sanity checks should not be included.
Selected metrics are exported to be measured, please review if these are enough.

More test configs are to be added later, future diffs should also include new test config if current configs can't cover the change.

Reviewed By: therealgymmy

Differential Revision: D29072997

fbshipit-source-id: ac15e763c1f0f461986f240cffd7acd6d40033f1