Skip counting check sum errors due to false positives
Summary:
In Bighash, we track checksum errors. However, we don't know if the bucket is
initialized or not when we read. If the checksum matches, we assume it was
initialized. If the check sum does not match, we can't differentiate between a
previously un-written bucket vs bad device.

to circumvent this, in BigHash code, we only bump up checksum errors when the
bloom filter tells the bucket has a key. However, this can have false
positives.

In short, we need another approach to reliably track check sum errors in
BigHash and not over-count un-initialized buckets as check sum errors.

Reviewed By: haowu14

Differential Revision: D29246343

fbshipit-source-id: d804444c2c6443f537051b61b73971dd4241f71c