Hold TombStone until remove job complete, drop FillLock in remove.
Summary:
Get FillLock and CancelFillMap are unnecessary when TombStone is valid.
We will hold the tombstone until the remove job completed, the tombstone will prevent any concurrent get and put jobs to be enqueued. Any immediate put job after remove (include insertOrReplace) will fail, this is fine in production since item won't be evicted so quickly, but we need to take care of unit tests where we usually force push to nvm after insertion.

TombStone now is required for all Nvm::remove.

This also helps to avoid a dead lock in D27834115.

Reviewed By: sathyaphoenix

Differential Revision: D28886405

fbshipit-source-id: ecbeb7d656ff18d41f41b64024a8233f203cd1f6