Fix config names affected by cachelib oss cleanup
Summary:
Recently, we cleaned up a bunch of cachebench configs. We deleted unused ones and renamed the ones with "dipper" into "navy".

This diff fixes those configs in run_simulator.
I do this by making sure that everything with "dipper" is fixed in run_simulator.py. I also went through the stack of diffs in D29719639 to make sure things are covered.

Reviewed By: therealgymmy

Differential Revision: D29856346

fbshipit-source-id: 420b8abec0b0acef70c39771dfda228d6aa90527