Delete HitsReinsertionPolicy recover API and its callsite
Summary:
We previously stored "item hits" information in the policy, and needed to persist and recover those across restarts.
Now we embed item hits directly in the index and no longer need to persist/recover for the policy.

In this diff, we will only remove recover() API and keep persist() logic for now. We plan to remove persist() API after BigCache has rolled out the release that no longer tries to recover the persistence policy.

Reviewed By: therealgymmy

Differential Revision: D29720069

fbshipit-source-id: 3b5c3c97256783ea03e148900ab38461a7d6fbe1