Fix broken test "ReclaimCorruptionSizeClass"
Summary:
Client Impact: none
----
This test was broken because I expected a header checksum failure, but when I mocked the write, I skipped the last 500 bytes. This means the entire header was skipped, so the checksum of the header is zero which is "correct". The way to fix this is to just skip the lat 10 bytes, which is much smaller than the entire header. This ensures the header checksum is invalid.

Reviewed By: sathyaphoenix

Differential Revision: D30113783

fbshipit-source-id: 5e5af8f81355d8ccd5db7a2cffa667a83a4e2014