Fix flush-forever-loop on device error for BlockCache in-mem buffer mode
Summary:
**Context**
Navy has a write mode "In-mem buffer mode" that buffers individual items into an entire region and then flush it to the device. However, if the device returned an error, we will retry indefinitely until success. This is an issue that if the device always returns an error, Navy cannot make any progress but instead will retry flushing the buffer again and again.
------

This diff:
- add a configurable retry limit in `BlockCache::Config`
- add 2 new stats to track flush failure and cleanup retries
- clean up buffer on flush failure
- split flush work and buffer detach work
- update device failure unit tests
- add unit tests to test new flush logic

Reviewed By: therealgymmy

Differential Revision: D29607292

fbshipit-source-id: 0dbddaba7bf10d1e9bc4d0633fe0d124fb496363