Prep for folly::Singleton in initFacebook (9/9) - Add folly::Singleton handling for fork() in Cachelib allocator test
Summary: `initFacebook` has a new THRIFT_FLAG evaluation as of D30235560 (top of this stack). This addition has the side effect of creating a `folly::Singleton`. Singletons must be handled (usually destroyed) prior to forking for safety reasons. This diff adds proper pre-fork Singleton handling to a Cachelib allocator test that performs a fork().

Reviewed By: mshneer

Differential Revision: D30250900

fbshipit-source-id: c148caf84cecaab5a2341fd0eb2cd79fc37be63b