Rate limit reclaiming when free/rss memory is decreasing/increasing to prevent OOM
Summary:
When a cachelib application warm restarts, the heap has to be built back again. In the meantime all the advised memory may get reclaimed which leaves application vulnerable to OOMs if heap grows rapidly.
This change adds a rate limiter for reclaiming. The rate limiter tracks free/rss memory over a window and computes rate of free/rss memory change based on values at the start and end of the window. When reclaiming, we assume that the memory being reclaimed in a given iteration would be all reclaimed within the poll interval to calculate reclaim rate. The difference of the two rate determines the amount of reclaimed memory. This effectively throttles reclaim when free/rss memory is falling/rising preventing OOM.

Reviewed By: therealgymmy

Differential Revision: D30852405

fbshipit-source-id: ad16cfaa4aec21b35befdb90aaf5d6ae7c7f6cec