Remove index from ReinsertionPolicy
Summary:
navy::Index is only used in HitReinsertionPolicy and what we need is just the LookupResult, which is always ready before we call the reinsertion policy.
This diff makes policy.shouldReinsert takes both HashedKey and LookupResult. I think it makes sense because the policy may want to depend on either the key or value to make a decision.

Reviewed By: sathyaphoenix

Differential Revision: D31396766

fbshipit-source-id: 1f9d9b44c325508dc8567883412003541c39e4de