Update the definition of compatibility
Summary:
I want to define compatible as a change that wouldn't cause compilation failure (because we didn't remove any public API) and wouldn't require dropping the cache, when someone is updating from a previous version.

And I want to declare that when an incompatible change is made, the version number will be bumped.

One thing that hangs for me is whether we have a clear definition of "public API". Is it everything under cachelib/allocator? Should we create a public interface class and call those public API?

Reviewed By: therealgymmy

Differential Revision: D31511563

fbshipit-source-id: 49a5a568ae30e248f4008e65a8d1081e00f35d79