resolve allocator corruption when requested mem is larger that 4mb
Summary:
# Context
after turning on forwarding writes from stateless writer to cache service for unified interest, cache service started to crash due to memory corruption. this turned out to be due to our prefix range (represented as F14Map) to try to allocate contiguous memory of larger than 4MB which is the limit on cachelib (until V2 is ready) that was causing memory corruption. To fix this, we add check in cachelib and we are also moving to use std::unordered_map to represent our ranges which does not enforce allocation of contiguous memory.

# This diff includes
- add check to cachelib to throw when allocation of more than 4MB is requested.
- remove std::move when setting items into cache (prefix/range scan) , as the actual operation is a copy and we are not doing any move underneath.
- add cachelib memory monitor config to add protection against OOMs.

Differential Revision: D31488495

fbshipit-source-id: 6657d62908bf3ef47ee35bedfa7a2637f9735786