Construct HitsReinsertionPolicy with index
Summary:
The goal of this stack of diffs is to find a clean way for CacheLib clients to create a customized block cache reinsertion policy.
Instead of simply expose the existing ReinsertionPolicy interface, this stack of diff removes the dependencies from ReinsertionPolicy to block_cache internals, and moves it to a more accessible location (cachelib/allocator/nvmcache).
And the motivation for this goal is that BigCache is going to try out ML based reinsertion policy on edge.
------

In this diff, create HitsReinsertionPolicy with index, and change the ReinsertionPolicy to only take a StringPiece.

Reviewed By: sathyaphoenix

Differential Revision: D31500253

fbshipit-source-id: 97c4e4f12c9f5d7fe5791df20339c5fd83364cd1