Enforce flush retry limit only on device failure
Summary:
Client Impact: Retry limit will be enforced only due to device failure

----

Previously, we enforce retry limit if there're still writers pending. What this means is that some times when writes are queued up, we end up rescheduling more than retry limit and still cannot flush the region. By this point, we will fail the flush.

We have observed this happen on memcache TFH region.

This diff changes the logic to ignore the retry limit if we're rescheduling due to pending writes.

However, a better change is to figure out how we can avoid being rescheduled due to pending writes. One thought is to let the last writer trigger the flush instead of the last allocation.

Reviewed By: leozzx

Differential Revision: D32088470

fbshipit-source-id: f107f51334a6d3a1f5ffe35591edefd2fd86bb9e