Move ReinsertionPolicy to cachelib/allocator/nvmcache
Summary:
The goal of this stack of diffs is to find a clean way for CacheLib clients to create a customized block cache reinsertion policy.
Instead of simply expose the existing ReinsertionPolicy interface, this stack of diff removes the dependencies from ReinsertionPolicy to block_cache internals, and moves it to a more accessible location (cachelib/allocator/nvmcache).
And the motivation for this goal is that BigCache is going to try out ML based reinsertion policy on edge.
----
Move ReinsertionPolicy into cachelib/allocator/nvmcache and rename to BlockCacheReinsertionPolicy.

Reviewed By: therealgymmy

Differential Revision: D31506764

fbshipit-source-id: f90f3a8673c9c5dfee3111e37419f8a8130096c5