skip promoting children items if promoting parent is not success
Summary:
For chained items access, we promote the parent and all children items in the chain. This could spend some CPU util when the chain is very long (e.g. 30 children on average, takes 10% cpu).
Even when the updateOnRead is disabled, cpu util is not reduce since we still iterate every child. It is not necessary to iterate the whole chain when the parent item is not promoted.

Reviewed By: sathyaphoenix

Differential Revision: D31798732

fbshipit-source-id: d5d1d53e62738d5e7e5babfcefd7b0ccfd267965