portability: add JSONConfig::SetVal for size_t if different from uint64_t
Summary:
JSONConfig::SetValImpl<TYPENAME> is implemented for uint64_t.
For systems where size_t is not the same as uint64_t,
add explicit implementation for it.
Use C++'s SFINAE ("Substitution Failure Is Not An Error")
to generate it only if size_t is different (to avoid compilation errors).
See:
  https://en.cppreference.com/w/cpp/language/sfinae
  https://techoverflow.net/2019/06/13/stdenable_if-and-stdis_same-minimal-example/
  https://eli.thegreenplace.net/2014/sfinae-and-enable_if/

Reviewed By: therealgymmy

Differential Revision: D33742171

fbshipit-source-id: cc7400084bdd267f24b00632826800c0809e00b7