Make preciseRemove an option for BlockCache
Summary:
As we discussed in this doc: https://fb.quip.com/b3GVAcKgzMqR.
We want to make an option for cachelib user to enable preciseRemove: checking the entire key when doing a remove, instead of only the hash value.

This remove is NvmCache::remove, it includes remove that comes from the CacheLib API as well as remove triggered internally such as allocation failure and device failure.
Turning this on will result a lookup when remove.

-----

I'll have another diff to let the overwrite in index triggers the itemDestructor.

Reviewed By: therealgymmy

Differential Revision: D34332620

fbshipit-source-id: 2cc3e0144cd7a3ec6bc78fe91bf6733a31fc0601