implemented a feature to detect the slab rebalancer stuck
Summary:
This commit implements a feature to detect if the slab rebalancer is stuck. Specifically, this
change declares the slab rebalancer is stuck if it has not made progress for a certain threshold
amount of time in releasing a given slab release context. This change adds a new
CacheAllocatorConfig option slabReleaseStuckThreshold for specifying the threshold. The stuck status
is exported via ServiceDataExporter as well. This change also adds a unit test to validate the
logic.

Reviewed By: therealgymmy

Differential Revision: D36641665

fbshipit-source-id: 9f8a3f477970da90f0d9ce8066320321add32afb