Allow user to specify which queue's age LruTailAgeStrategy will use to rebalance
Summary:
Client Impact: users can choose which queue's eviction age to rebalance

----
In LruTailAgeStrategy, we check for eviction-age by calling “getOldestElementAge()” (https://fburl.com/code/bbd51ve1). That function relies on the “warm-queue age”. For LRU, this is the same as the queue’s eviction-age. For 2Q, it refers to warm-queue. But in memcache with 2Q (switched over in D35137292), we have disabled the warm-queue to retain LRU promotion behavior. So, this now always return 0, and rebalancer will NOT rebalance correctly.

This should address S274222

Reviewed By: alikhtarov, stuclar

Differential Revision: D36882849

fbshipit-source-id: 91b128e664075e706148d79c6cc5217d5f7111f7