Fix test "NvmCacheTest.NvmClean"
Summary:
Previously this test didn't flush nvm-cache while evictions are happening. What would have happened is that some evictions from DRAM didn't make their way into NvmCache due to tombstone. The reason is because a remove job to NVM was created when a key was first inserted into DRAM cache. And then when this key is evicted into NVM, the remove job has not yet completed. This practically cannot happen in production but in test where an item stays in cache maybe only for milliseconds, this race can happen.

This test relies on all items being successfully evicted to NvmCache. So we need to ensure all Nvm Puts succeed. The fix is to periodically flush NvmCache to make sure a remove job doesn't linger around for too long.

Reviewed By: jiayuebao

Differential Revision: D36994802

fbshipit-source-id: dbe3af43d3224459e894ff8f693af343fcca8e05